requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-limiter==3.11.0
gradio-client==1.3.0
Pillow==11.1.0
python-dotenv==1.1.0
gunicorn==23.0.0
opencv-python-headless==4.9.0.80
numpy<2
